+++ osrframework/wrappers/askfm.py
@@ -97,7 +97,8 @@ class Askfm(Platform):
         # Definition of regular expressions to be searched in usufy mode
         self.fieldsRegExp["usufy"] = {}
         # Example of fields:
-        #self.fieldsRegExp["usufy"]["i3visio.location"] = ""
+        self.fieldsRegExp["usufy"]["i3visio.fullname"] = {"start": 'class="blockLink">', "end": '</span>'}
+        self.fieldsRegExp["usufy"]["i3visio.uri_image_profile"] = {"start": '<link rel="image_src" href="', "end": '" />'}
         
         # Definition of regular expressions to be searched in searchfy mode
         #self.fieldsRegExp["searchfy"] = {}