+++ doc/CHANGES.md
@@ -12,6 +12,7 @@ For more information, check the README.md in <http://github.com/i3visio/osrframe
 - Add feature: Provide an API to be used by other applications
 - Add feature: Provide a GUI to make it easier to launch the commands for newcomers
 - Add feature: Change the path where Maltego settings file is created
+- Add feature: Add metadata extraction for all usufy wrappers starting with 'a'
 - Add feature: Add Affilorama as a new usufy platform
 - Add feature: Add Abou as a new usufy platform
 - Add feature: Add Canva as a new usufy platform